refactor: unify shared types into @vibe-replay/types package
Extract Scene, Annotation, DataSourceInfo, and ReplaySession into a
shared `packages/types` package to eliminate manual sync between
CLI and viewer type definitions. Both packages now re-export from
the single source of truth.
