+++ packages/cli/package.json
@@ -41,6 +41,7 @@
     "test": "vitest run"
   },
   "dependencies": {
+    "@vibe-replay/types": "workspace:*",
     "@hono/node-server": "^1.19.11",
     "@inquirer/prompts": "^7.2.0",
     "chalk": "^5.3.0",
