+++ packages/cli/src/types.ts
@@ -1,3 +1,8 @@
+// Shared types — single source of truth
+export type { Annotation, DataSourceInfo, ReplaySession, Scene } from "@vibe-replay/types";
+
+// CLI-only types below
+
 export interface SessionInfo {
   provider: string;
   sessionId: string;
@@ -71,79 +76,3 @@ export type ContentBlock =
 export type ToolResultContent =
   | { type: "text"; text: string }
   | { type: "tool_result"; tool_use_id: string; content: string };
-
-export type Scene =
-  | { type: "user-prompt"; content: string; timestamp?: string; images?: string[] }
-  | { type: "compaction-summary"; content: string; timestamp?: string }
-  | { type: "thinking"; content: string; timestamp?: string }
-  | { type: "text-response"; content: string; timestamp?: string }
-  | {
-      type: "tool-call";
-      toolName: string;
-      input: Record<string, any>;
-      result: string;
-      timestamp?: string;
-      diff?: { filePath: string; oldContent: string; newContent: string };
-      bashOutput?: { command: string; stdout: string };
-      images?: string[];
-    };
-
-export interface Annotation {
-  id: string;
-  sceneIndex: number;
-  selectedText?: string;
-  body: string;
-  author: string;
-  createdAt: string;
-  updatedAt: string;
-  resolved: boolean;
-}
-
-export interface DataSourceInfo {
-  primary: string;
-  sources: string[];
-  supplements?: string[];
-  notes?: string[];
-}
-
-export interface ReplaySession {
-  meta: {
-    sessionId: string;
-    slug: string;
-    title?: string;
-    provider: string;
-    dataSource?: string;
-    dataSourceInfo?: DataSourceInfo;
-    startTime: string;
-    endTime?: string;
-    model?: string;
-    cwd: string;
-    project: string;
-    generator?: {
-      name: string;
-      version: string;
-      generatedAt: string;
-    };
-    stats: {
-      sceneCount: number;
-      userPrompts: number;
-      toolCalls: number;
-      thinkingBlocks?: number;
-      durationMs?: number;
-      tokenUsage?: {
-        inputTokens: number;
-        outputTokens: number;
-        cacheCreationTokens: number;
-        cacheReadTokens: number;
-      };
-      costEstimate?: number;
-    };
-    compactions?: Array<{
-      timestamp: string;
-      trigger: string;
-      preTokens?: number;
-    }>;
-  };
-  scenes: Scene[];
-  annotations?: Annotation[];
-}
