+++ packages/types/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@vibe-replay/types",
+  "version": "0.0.1",
+  "private": true,
+  "type": "module",
+  "license": "MIT",
+  "exports": {
+    ".": {
+      "types": "./src/index.ts",
+      "import": "./src/index.ts"
+    }
+  },
+  "files": [
+    "src"
+  ]
+}
